Enhancements to I18n handling when dealing with enumerations
A couple of changes around this:

  * Modified the backend to support the same method for looking up
    enumeration values (which might contain dots and other interesting
    characters).  Pulled the code that handles this out from the
    frontend initializer and into a common spot that both backend and
    frontend can use.

  * Have the backend load any plugin-supplied I18n yml files.
    Previously it was only looking at the distribution defaults.

  * Modify aspace_form_helper.rb to use the custom enumeration syntax
    when resolving enumeration values.  Ensures that enumeration values
    containing dots still get translated correctly.
